Added a timer and changed the shooting projectile
Will be working on different affects for the "bullets"(rubber bands). As well as the action for the bombs. The action for the bomb could be that you collect and object. With  the object in the player_manager the player can drop the bomb one per every 10 seconds. This could be a concept.